Rewrite Slider
